pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "procureme"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "MrDataPsycho", email = "mr.data.psycho@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "sqlmodel==0.0.21",
  "psycopg2-binary==2.9.10",
  "pandas",
  "lancedb==0.22.0",
  "fastapi==0.115.12",
  "uvicorn==0.34.2",
  "pydantic==2.11.4",
  "streamlit==1.45.1",
  "watchdog==6.0.0",
  "openai==1.82.1",
  "pydantic-settings"
]

[project.optional-dependencies]
testing = [
    "pytest",
    "pytest-cov",
    "pytest-mock"
]

etl = [
  "pymupdf4llm==0.0.24",
  "doctomarkdown==0.1.7",
  "neopipe==0.1.1",
  "ollama==0.4.8",
  "llama-index==0.12.36",
  "llama-index-llms-ollama==0.5.4",
  "llama-index-embeddings-ollama==0.6.0",
]

dev = [
  "ruff",
  "mypy",
  "typer",
  "python-dotenv",
  "polars",
  "nbconvert",
  "requests",
  "ipywidgets",
  "typer",
  "tqdm",
  "aiofiles",
  "tenacity",
  "httpx"
]

[project.urls]
Documentation = "https://github.com/MrDataPsycho/smart-procurement#readme"
Issues = "https://github.com/MrDataPsycho/smart-procurement/issues"
Source = "https://github.com/MrDataPsycho/smart-procurement"

[tool.hatch.version]
path = "src/procureme/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0"
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/procureme tests}"

[tool.coverage.run]
source_pkgs = ["procureme", "tests"]
branch = true
parallel = true
omit = [
  "src/procureme/__about__.py",
]

[tool.coverage.paths]
procureme = ["src/procureme", "*/smart-procurement/src/procureme"]
tests = ["tests", "*/smart-procurement/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.hatch.envs.docs]
dependencies = [
  "mkdocs"
]
[tool.hatch.envs.docs.scripts]
build = "mkdocs build --clean --strict"
serve = "mkdocs serve --dev-addr localhost:8000"
